se implementan gates para la autorizacion

+++ app/Http/Controllers/QuestionsController.php
@@ -62,6 +62,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question )){
+            abort(403, "access denied");
+        }
+        
         return view('questions.edit', compact('question'));
     }
 
@@ -74,6 +78,9 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question )){
+            abort(403, "access denied");
+        }
         $question->update($request->only('title','body'));
         return redirect()->route('questions.index')->with('success', "your question has been updated");
     }
@@ -86,6 +93,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question )){
+            abort(403, "access denied");
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success', "your question has been deleted");
     }
